add poetry, linting
